refactor: orthography part of word_metadata.rs
