+++ harper-core/src/lib.rs
@@ -27,6 +27,7 @@ mod token_kind;
 mod token_string_ext;
 mod vec_ext;
 pub mod word_metadata;
+pub mod word_metadata_orthography;
 
 use render_markdown::render_markdown;
 use std::collections::VecDeque;
